refactor: 진행 중인 도장판 컴포넌트화 및 VAC 패턴 적용
